Bugfixes, revamped language system
Fixed a bug relating to the player voice distance where the input by the player wasn't squared on the server (and the other way around in the editor). Also revamped the language system to be MUCH more intuitive. All that is needed is for someone to make a copy of the english lang file, edit to their needs, and restart the map/server. The language will immediately be an option in the editor and will fall back to English if a given phrase is not found.